extract snapshot client, auth and payload
